Game is just about playable (game states not implemented yet ofc), but needs some tweaking with movement speed, etc. when there are multiple clients

+++ src/Game/PlayerObjects/Vec2D.java
@@ -2,13 +2,13 @@ package Game.PlayerObjects;
 
 public class Vec2D {
     public Coord coord;
-    private int angle;
+    private double angle;
 
-    public void setAngle(int a) {
+    public void setAngle(double a) {
         angle = a;
     }
 
-    public int getAngle() {
+    public double getAngle() {
         return angle;
     }
